+++ packages/ts-sdk/src/functions/keyless.ts
@@ -9,6 +9,7 @@
 // Standalone functions
 export {
   getPepper,
+  getPepperAndAddress,
   getProof,
   deriveKeylessAccount,
   updateFederatedKeylessJwkSetTransaction,